requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.3
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
